chore: move AppMap webview sources to new package
