+++ plugin-core/src/main/java/appland/webviews/appMap/AppMapNotificationProvider.java
@@ -1,4 +1,4 @@
-package appland.editor;
+package appland.webviews.appMap;
 
 import appland.AppMapBundle;
 import appland.files.AppMapFiles;
